+++ lib/torrent_search/controllers/search.rb
